update the reset --quiet path codepath to pass the correct flags to the post-indexchanged hook
In the reset --mixed code path, the index is created from scratch from the
given commit by the call to read_from_tree().  Since this is the code that
actually modifies the index, make sure we set the the_index.updated_skipworktree
flag which is passed to the post-indexchanged hook.

Updated the post-index-changed-hook test script to pass the --quiet flag
so that we can prevent future regressions.

Signed-off-by: Ben Peart <benpeart@microsoft.com>